+++ BeSafeWebApp.Contracts/Interfaces/Repositories/ICatagoryRepository.cs
@@ -7,7 +7,7 @@ namespace BeSafeWebApp.Contracts.Interfaces
     public interface ICategoryRepository : IGenericRepository<Entities.Category>
     {
         Task<IList<Entities.Category>> GetAllCategories();
-        Task<Entities.Category> GetCategoryById(int id);
+        Task<Entities.Category> GetCategoryById(long id);
 
         //Task<int> AddUser(Entities.User inputEt);
         //Task UpdateUser(Entities.User inputEt);
